fix: receiveAttack() method not work properly
Ensure receiveAttack() method returns value when hitting the ship target
position.

Before, the receiveAttack() method did not return a value when hitting
the ship's target position, causing styleCell() to fail to add the
correct class to the clicked cell.

Now, receiveAttack() returns the appropriate value when hitting the
ship's target position, allowing styleCell() to add the 'hit' class to
the clicked cell.